refactor to make the common code be a setup function
This is so that we can make both a Node.js instance and
web browser instance for when `--inspect` is used in Node.js.
